Clean up for building on linux

+++ Editor/src/ImGui/Viewers/ImGuiTextureView.cpp
@@ -18,7 +18,7 @@ void ImGuiTextureView::OnImGuiRender()
 
 	if (ImGui::Begin(m_WindowName.c_str(), m_Show))
 	{
-		ImGui::Image(m_Texture, ImVec2(m_Texture->GetWidth(), m_Texture->GetHeight()));
+		ImGui::Image(m_Texture, ImVec2((float)m_Texture->GetWidth(), (float)m_Texture->GetHeight()));
 	}
 	ImGui::End();
 }
